wip: running state

+++ test/conftest.py
@@ -16,8 +16,10 @@ def pytest_sessionstart(session):
     Called after the Session object has been created and
     before performing collection and entering the run test loop.
     """
+    # remove the tmp folder to empty old results
     p = Path(__file__).parent.parent / 'tmp'
-    shutil.rmtree(p)
+    if p.exists() and p.is_dir():
+        shutil.rmtree(p)
     p.mkdir(exist_ok=True)
